refactor(javascript): move patch_url_to_point_to_proxy to package_managers.general
The function has no npm-specific logic, and bundler needs it too for
gem downloads.
