+++ hermeto/core/package_managers/javascript/npm/__init__.py
@@ -1,9 +1,6 @@
 # SPDX-License-Identifier: GPL-3.0-only
 from hermeto.core.package_managers.javascript.npm.main import fetch_npm_source
-from hermeto.core.package_managers.javascript.npm.resolver import (
-    async_download_with_auth,
-    patch_url_to_point_to_proxy,
-)
+from hermeto.core.package_managers.javascript.npm.resolver import async_download_with_auth
 from hermeto.core.package_managers.javascript.npm.utils import (
     NPM_REGISTRY_URL,
     YARN_REGISTRY_URL,
@@ -16,5 +13,4 @@ __all__ = [
     "YARN_REGISTRY_URL",
     "is_from_npm_registry",
     "async_download_with_auth",
-    "patch_url_to_point_to_proxy",
 ]
